fixed strange testing error
